pulled out request token to its own object. added verifyToken() call to httpserver getexchange handler with a todo
