+++ httpclient/src/test/kotlin/tbdex/sdk/httpclient/RequestTokenTest.kt
@@ -1,5 +1,7 @@
 package tbdex.sdk.httpclient
 
+import assertk.assertThat
+import assertk.assertions.containsExactlyInAnyOrder
 import com.nimbusds.jwt.SignedJWT
 import web5.sdk.crypto.InMemoryKeyManager
 import web5.sdk.dids.methods.dht.DidDht
@@ -8,14 +10,14 @@ import kotlin.test.assertEquals
 import kotlin.test.assertNotNull
 import kotlin.test.assertTrue
 
-class UtilsTest {
+class RequestTokenTest {
 
   @Test
   fun `generateRequestToken() generates a JWT`() {
     val did = DidDht.create(InMemoryKeyManager())
     val pfiDid = "did:ion:123"
 
-    val token = generateRequestToken(did, pfiDid)
+    val token = RequestToken.generateRequestToken(did, pfiDid)
     assertNotNull(token)
   }
 
@@ -24,13 +26,11 @@ class UtilsTest {
     val did = DidDht.create(InMemoryKeyManager())
     val pfiDid = "did:ion:123"
 
-    val token = generateRequestToken(did, pfiDid)
+    val token = RequestToken.generateRequestToken(did, pfiDid)
     val claimsSet = SignedJWT.parse(token).jwtClaimsSet
 
-    val requiredKeys = listOf("aud", "iss", "exp", "jti", "iat")
-    requiredKeys.forEach {
-      assertTrue(claimsSet.claims.containsKey(it))
-    }
+    assertThat(claimsSet.claims.keys)
+      .containsExactlyInAnyOrder(RequestToken.requiredClaimKeys)
   }
 
   @Test
@@ -38,7 +38,7 @@ class UtilsTest {
     val did = DidDht.create(InMemoryKeyManager())
     val pfiDid = "did:ion:123"
 
-    val token = generateRequestToken(did, pfiDid)
+    val token = RequestToken.generateRequestToken(did, pfiDid)
     val claimsSet = SignedJWT.parse(token).jwtClaimsSet
 
     assertTrue(claimsSet.issuer.contains(did.uri))
@@ -51,9 +51,9 @@ class UtilsTest {
     val did = DidDht.create(InMemoryKeyManager())
     val pfiDid = "did:ion:123"
 
-    val token = generateRequestToken(did, pfiDid)
+    val token = RequestToken.generateRequestToken(did, pfiDid)
 
-    val verificationResult = verifyRequestToken(token, pfiDid)
+    val verificationResult = RequestToken.verifyRequestToken(token, pfiDid)
 
     assertEquals(did.uri, verificationResult)
   }
