+++ httpserver/src/main/kotlin/tbdex/sdk/httpserver/handlers/GetExchanges.kt
@@ -4,6 +4,7 @@ import io.ktor.http.HttpHeaders
 import io.ktor.http.HttpStatusCode
 import io.ktor.server.application.ApplicationCall
 import io.ktor.server.response.respond
+import tbdex.sdk.httpclient.RequestToken
 import tbdex.sdk.httpclient.models.ErrorDetail
 import tbdex.sdk.httpserver.models.ErrorResponse
 import tbdex.sdk.httpserver.models.ExchangesApi
@@ -62,11 +63,9 @@ suspend fun getExchanges(
     return
   }
 
-  // todo: verify JWT token using new CryptoUtils.verify() method
-  // to be written to address these issues:
-  // generating JWT token: https://github.com/TBD54566975/tbdex-kt/issues/121
-  // verifying JWT token: https://github.com/TBD54566975/tbdex/issues/210
-
+  val token = arr[1]
+  // TODO: how to access pfiDid here?
+  RequestToken.verifyRequestToken(token, "")
   val exchanges = exchangesApi.getExchanges()
 
   if (callback != null) {
